+++ src/Components/App/App.js
@@ -1,9 +1,12 @@
 import './App.css';
+import React from 'react';
 import {BrowserRouter, Route, Switch, NavLink} from 'react-router-dom';
-import logo from '../Assets/lyricalLiesLogo.png'
-import faves from '../Assets/favesIcon.png'
-import home from '../Assets/homeIcon.png'
-function App() {
+import logo from '../../Assets/lyricalLiesLogo.png'
+import faves from '../../Assets/favesIcon.png'
+import home from '../../Assets/homeIcon.png'
+import LandingPage from '../LandingPage/LandingPage.js'
+
+const App = () => {
   return (
     <BrowserRouter>
       <main>
@@ -19,7 +22,7 @@ function App() {
             <NavLink id="homeButton" activeClassName="activeHomeButton" to='/home'>
               <img
                 className="navImg"
-                title="Return To Lobby"
+                title="Return to Homepage"
                 src={home}
                 alt="Navigate back to home page"
               />
@@ -29,15 +32,20 @@ function App() {
             <NavLink id="favesButton" activeClassName="activeHomeButton" to='/faves'>
               <img
                 className="navImg"
-                title="Return To Lobby"
+                title="View Favorite Lyrics"
                 src={faves}
-                alt="Navigate back to home page"
+                alt="Navigate to favorites page"
               />
             </NavLink>
           </Route>
           </section>
         </header>
       <Switch>
+        <Route path='/home'></Route>
+        <Route path='/lyrics'></Route>
+        <Route exactPath='/'>
+          <LandingPage />
+        </Route>
       </Switch>
       </main>
     </BrowserRouter>
